refactor(common): CHECKOUT-4338 Move `BuildHooks` Webpack plugin to `scripts` folder
